"updated skip_list impln with ignoring the first n elements from prefetching + different loop configurations (pipeline and unroll)"

+++ HLS_proj/proj_12_Linkedlist_prefetch/LL_prefetch_test.cpp
@@ -3,7 +3,7 @@
 const int n = 10000;
 
 typedef int dat_typ;
-typedef int dat_typ1;
+typedef short dat_typ1;
 
 #define level 2
 #define skip_intr 20
